+++ Server/ClientChatWidget.cpp
@@ -3,24 +3,22 @@
 
 ClientChatWidget::ClientChatWidget(QTcpSocket *client, QWidget *parent) :
     QWidget(parent),
-    ui(new Ui::ClientChatWidget),
-    _client(client)
-
+    ui(new Ui::ClientChatWidget)
 {
     ui->setupUi(this);
-    connect(_client, &QTcpSocket::readyRead, this, &ClientChatWidget::dataReceived);
-    connect(_client, &QTcpSocket::disconnected, this, &ClientChatWidget::clientDisconnected);
+    _client = new ClientManager(client, this);
+//    connect(_client, &QTcpSocket::readyRead, this, &ClientChatWidget::dataReceived);
+    connect(_client, &ClientManager::disconnected, this, &ClientChatWidget::clientDisconnected);
+    connect(_client, &ClientManager::textMessageReceived, this, &ClientChatWidget::textMessageReceived);
+    connect(_client, &ClientManager::isTyping, this, &ClientChatWidget::onTyping);
+    connect(_client, &ClientManager::nameChanged, this, &ClientChatWidget::clientNameChanged);
+    connect(_client, &ClientManager::statusChanged, this, &ClientChatWidget::statusChanged);
+    connect(ui->lnMessage, &QLineEdit::textChanged, _client, &ClientManager::sendIsTyping);
 }
 
 ClientChatWidget::~ClientChatWidget()
 {
     delete ui;
-}
-
-void ClientChatWidget::dataReceived()
-{
-    auto data = _client->readAll();
-    ui->lstMessages->addItem(data);
 }
 
 void ClientChatWidget::clientDisconnected()
@@ -30,7 +28,19 @@ void ClientChatWidget::clientDisconnected()
 
 void ClientChatWidget::on_btnSend_clicked()
 {
-    _client->write(ui->lnMessage->text().trimmed().toUtf8());
+    auto message = ui->lnMessage->text().trimmed();
+    _client->sendMessage(message);
     ui->lnMessage->setText("");
+    ui->lstMessages->addItem(message);
+}
+
+void ClientChatWidget::textMessageReceived(QString message)
+{
+    ui->lstMessages->addItem(message);
+}
+
+void ClientChatWidget::onTyping()
+{
+    emit isTyping(QString("%1 is typing...").arg(_client->name()));
 }