+++ Server/ClientChatWidget.h
@@ -1,6 +1,8 @@
 #ifndef CLIENTCHATWIDGET_H
 #define CLIENTCHATWIDGET_H
 
+#include "ClientManager.h"
+
 #include <QTcpSocket>
 #include <QWidget>
 
@@ -16,14 +18,22 @@ public:
     explicit ClientChatWidget(QTcpSocket *cleint, QWidget *parent = nullptr);
     ~ClientChatWidget();
 private slots:
-    void dataReceived();
+
     void clientDisconnected();
 
     void on_btnSend_clicked();
 
+    void textMessageReceived(QString message);
+    void onTyping();
+
+signals:
+    void clientNameChanged(QString name);
+    void isTyping(QString message);
+    void statusChanged(ChatProtocol::Status status);
+
 private:
     Ui::ClientChatWidget *ui;
-    QTcpSocket *_client;
+    ClientManager *_client;
 };
 
 #endif // CLIENTCHATWIDGET_H