+++ CHANGELOG.md
@@ -6,3 +6,12 @@
 - Simple Chat Client/Server apps created
     - Sending text messages
     dynamic UI in server based on client connecting
+
+## v2.0.0
+
+### Added or Changed
+- Add following options to chat applications:
+    - Send Status
+    - Send Client Name
+    - Send IsTyping Message
+